sim/LosHandler: remove CREG
It isn't used in this part of the engine and it creates some complications
when inverting relation between LosHandler and its dependencies by passing
the dependencies to the CLosHandler constructor.

Even if save/load using CREG is desired at some point everything the
LosHandler knows can be regenerated from units and configuration loaded
from the save file.